.devcontainer/test-requirements.txt
pyyaml==6.0
pytest==7.1.2
pytest-testinfra==6.7.0
python-on-whales==0.46.0
